join NbInterface and NbOpParameters into NbOp
